+++ packages/templates/src/index.ts
@@ -29,3 +29,4 @@ export * from './calc';
 export * from './impress';
 export * from './base';
 export * from './code-stamps/code-stamps';
+export * from './layout-stamps/layout-stamps';